DCAC-108: Reorganise code according to responsibilities.
